DeviceTree_templates: u-boot: correct st ethernet property
In this version of u-boot, the device tree node property
"st,eth-clk-sel" is not aligned with the kernel device tree and must
contains underscore instead of dash.
So we add this property with correct typo in the "-mx-u-boot.dtsi" files
dedicated to u-boot. Without it, ethernet does not work on u-boot with
our SOM.

Signed-off-by: Christophe Parant <c.parant@phytec.fr>