DeviceTree_templates: tf-a: remove memory node from user code
The "memory" node device tree (ddr size and address) is now generated
by CubeMX for the TF-A. So we remove it from user section.
Also the ddr size was wrong in this node for machine with 1GB RAM
(0x20000000 instead of 0x40000000).
But it seems that there is no  functionnal impact (this error is present
in upstream device tree).

Signed-off-by: Christophe Parant <c.parant@phytec.fr>